% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{doc_data}
\alias{doc_data}
\title{Document data frames}
\usage{
doc_data(d)
}
\arguments{
\item{d}{A data.frame}
}
\value{
An atomic character vector.
}
\description{
GivenFrom a data.frame, this function generates an atomic character vector
with skeleton documentation as per roxygen2's syntax.
}
\examples{
m <- data.frame(a = c(1, 2), b = c("n", "m"), stringsAsFactors = FALSE)
doc_data(m)

}

