% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_all.R
\name{ggsave_all}
\alias{ggsave_all}
\title{Save all ggplot objects}
\usage{
ggsave_all(dir = ".", devices = c("pdf", "png"), env = .GlobalEnv, ...)
}
\arguments{
\item{dir}{Directory path to save files.}

\item{devices}{A character vector of devices to use. Defaults to .pdf, but
can take any of the devices currently supported by \link[ggplot2]{ggsave}.
Specify multiple values to save versions of every plot using each device.}

\item{env}{The environment for scan for ggplot objects.}

\item{...}{Arguments passed to \link[ggplot2]{ggsave}.}
}
\description{
Saves all ggplot objects in the current environment to disk using \link[ggplot2]{ggsave}.
}

