% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_grid.R
\name{na_grid}
\alias{na_grid}
\title{Visualize missing values in a data frame}
\usage{
na_grid(d)
}
\arguments{
\item{d}{A data frame}
}
\description{
Checks for NAs in every column of a dataframe, visualizing them using ggplot's raster mode
}

