% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_per.R
\name{unique_per}
\alias{unique_per}
\title{Utility to check if a data frame column (name) has unique values based on a given unique ID column (uid)}
\usage{
unique_per(d, name, uid = names(d)[1])
}
\arguments{
\item{d}{Data frame}

\item{name}{Quoted column name}

\item{uid}{Quoted column name of unique id}
}
\description{
Utility to check if a data frame column (name) has unique values based on a given unique ID column (uid)
}

